- java is also working; - both cpp and java does not support batch yet.

+++ src/scag/pvss/pvap/PVAP.hpp
@@ -58,7 +58,7 @@ public:
   void decodeMessage(Serializer& ss)
   {
     uint32_t seqNum=ss.readInt32();
-    uint32_t tag=ss.readInt16();
+    uint16_t tag=ss.readInt16();
     switch(tag)
     {
       case tag_PC_DEL:
